5th part of FS#12176. Further fixed point migration. Only two emulators (ym2413, ym2612) still use floating point.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30281 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libgme/hes_apu_adpcm.h
@@ -43,7 +43,7 @@ struct Hes_Apu_Adpcm {
 
 	struct Blip_Buffer* output;
 	blip_time_t  last_time;
-	double       next_timer;
+	int          next_timer;
 	int          last_amp;
 };
