+++ apps/codecs/libgme/nes_vrc7_apu.c
@@ -29,7 +29,7 @@ void Vrc7_reset( struct Nes_Vrc7_Apu* this )
 	OPLL_setMask(&this->opll, this->mask);
 }
 
-void Vrc7_set_rate( struct Nes_Vrc7_Apu* this, double r )
+void Vrc7_set_rate( struct Nes_Vrc7_Apu* this, int r )
 {
 	OPLL_set_quality( &this->opll, r < 44100 ? 0 : 1 );
     OPLL_set_rate( &this->opll, (e_uint32)r );
